+++ test-packages/integration-tests/test/auth-flows/auth-flow.spec.ts
@@ -1,6 +1,6 @@
 import {
   executeHttpRequest,
-  fetchDestination,
+  fetchDestination, getDestination,
   getService,
   serviceToken,
   userApprovedServiceToken,
@@ -124,6 +124,16 @@ describe('OAuth flows', () => {
 
     expect(response.status).toBe(200);
   }, 60000);
+
+  xit('ClientCertificate: Fetches the certificate and uses it',async ()=>{
+    process.env.HTTPS_PROXY = 'http://someHost:1234';
+    process.env.NO_PROXY = 'https://s4sdk.authentication.sap.hana.ondemand.com/oauth/token';
+
+    const destination = await getDestination('CC8-HTTP-CERT');
+    expect(destination!.certificates!.length).toBe(1);
+    const bps = await BusinessPartner.requestBuilder().getAll().top(5).execute(destination!);
+    expect(bps.length).toBeGreaterThan(0);
+  },10000);
 
   xit('OAuth2UserTokenExchange: Subscriber destination and Subscriber Jwt', async () => {
     const subscriberDestToken = await serviceToken('destination', {